chore: Move Dialog component to UI directory
