--- packages/webapp-libs/webapp-core/src/components/dialog/dialog.component.tsx
@@ -1,5 +1,0 @@
-import * as DialogPrimitive from '@radix-ui/react-dialog';
-
-const Dialog = DialogPrimitive.Root;
-
-export { Dialog };
