--- packages/webapp-libs/webapp-core/src/components/dialog/dialogContent/index.ts
@@ -1,1 +1,0 @@
-export * from './dialogContent.component';
