--- packages/webapp-libs/webapp-core/src/components/dialog/dialogDescription/dialogDescription.component.tsx
@@ -1,14 +1,0 @@
-import * as DialogPrimitive from '@radix-ui/react-dialog';
-import { forwardRef } from 'react';
-
-import { cn } from '../../../lib/utils';
-
-const DialogDescription = forwardRef<
-  React.ElementRef<typeof DialogPrimitive.Description>,
-  React.ComponentPropsWithoutRef<typeof DialogPrimitive.Description>
->(({ className, ...props }, ref) => (
-  <DialogPrimitive.Description ref={ref} className={cn('text-sm text-muted-foreground', className)} {...props} />
-));
-DialogDescription.displayName = DialogPrimitive.Description.displayName;
-
-export { DialogDescription };
