--- packages/webapp-libs/webapp-core/src/components/dialog/dialogDescription/index.ts
@@ -1,1 +1,0 @@
-export * from './dialogDescription.component';
