--- packages/webapp-libs/webapp-core/src/components/dialog/dialogFooter/dialogFooter.component.tsx
@@ -1,8 +1,0 @@
-import { cn } from '../../../lib/utils';
-
-const DialogFooter = ({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) => (
-  <div className={cn('flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2', className)} {...props} />
-);
-DialogFooter.displayName = 'DialogFooter';
-
-export { DialogFooter };
