--- packages/webapp-libs/webapp-core/src/components/dialog/dialogFooter/index.ts
@@ -1,1 +1,0 @@
-export * from './dialogFooter.component';
