--- packages/webapp-libs/webapp-core/src/components/dialog/dialogHeader/dialogHeader.component.tsx
@@ -1,8 +1,0 @@
-import { cn } from '../../../lib/utils';
-
-const DialogHeader = ({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) => (
-  <div className={cn('flex flex-col space-y-1.5 text-center sm:text-left', className)} {...props} />
-);
-DialogHeader.displayName = 'DialogHeader';
-
-export { DialogHeader };
