--- packages/webapp-libs/webapp-core/src/components/dialog/dialogHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './dialogHeader.component';
