--- packages/webapp-libs/webapp-core/src/components/dialog/dialogOverlay/dialogOverlay.component.tsx
@@ -1,21 +1,0 @@
-import * as DialogPrimitive from '@radix-ui/react-dialog';
-import { forwardRef } from 'react';
-
-import { cn } from '../../../lib/utils';
-
-const DialogOverlay = forwardRef<
-  React.ElementRef<typeof DialogPrimitive.Overlay>,
-  React.ComponentPropsWithoutRef<typeof DialogPrimitive.Overlay>
->(({ className, ...props }, ref) => (
-  <DialogPrimitive.Overlay
-    ref={ref}
-    className={cn(
-      'fixed inset-0 z-50 bg-background/80 backdrop-blur-sm transition-all duration-100 data-[state=closed]:animate-out data-[state=closed]:fade-out data-[state=open]:fade-in',
-      className
-    )}
-    {...props}
-  />
-));
-DialogOverlay.displayName = DialogPrimitive.Overlay.displayName;
-
-export { DialogOverlay };
