--- packages/webapp-libs/webapp-core/src/components/dialog/dialogOverlay/index.ts
@@ -1,1 +1,0 @@
-export * from './dialogOverlay.component';
