--- packages/webapp-libs/webapp-core/src/components/dialog/dialogPortal/dialogPortal.component.tsx
@@ -1,10 +1,0 @@
-import * as DialogPrimitive from '@radix-ui/react-dialog';
-
-const DialogPortal = ({ children, ...props }: DialogPrimitive.DialogPortalProps) => (
-  <DialogPrimitive.Portal {...props}>
-    <div className="fixed inset-0 z-50 flex items-start justify-center sm:items-center">{children}</div>
-  </DialogPrimitive.Portal>
-);
-DialogPortal.displayName = DialogPrimitive.Portal.displayName;
-
-export { DialogPortal };
