--- packages/webapp-libs/webapp-core/src/components/dialog/dialogPortal/index.ts
@@ -1,1 +1,0 @@
-export * from './dialogPortal.component';
