--- packages/webapp-libs/webapp-core/src/components/dialog/dialogTitle/dialogTitle.component.tsx
@@ -1,18 +1,0 @@
-import * as DialogPrimitive from '@radix-ui/react-dialog';
-import { forwardRef } from 'react';
-
-import { cn } from '../../../lib/utils';
-
-const DialogTitle = forwardRef<
-  React.ElementRef<typeof DialogPrimitive.Title>,
-  React.ComponentPropsWithoutRef<typeof DialogPrimitive.Title>
->(({ className, ...props }, ref) => (
-  <DialogPrimitive.Title
-    ref={ref}
-    className={cn('text-lg font-semibold leading-none tracking-tight', className)}
-    {...props}
-  />
-));
-DialogTitle.displayName = DialogPrimitive.Title.displayName;
-
-export { DialogTitle };
