--- packages/webapp-libs/webapp-core/src/components/dialog/dialogTitle/index.ts
@@ -1,1 +1,0 @@
-export * from './dialogTitle.component';
