--- packages/webapp-libs/webapp-core/src/components/dialog/dialogTrigger/index.ts
@@ -1,1 +1,0 @@
-export * from './dialogTrigger.component';
