+++ packages/webapp-libs/webapp-core/src/components/ui/__tests__/dialog.component.spec.tsx
@@ -1,9 +1,9 @@
 import { fireEvent, screen } from '@testing-library/react';
 
-import { Dialog, DialogContent, DialogHeader } from '../';
 import { render } from '../../../tests/utils/rendering';
+import { Dialog, DialogContent, DialogHeader } from '../dialog';
 
-describe('Dialog: Component', () => {
+describe('UI/Dialog: Component', () => {
   const content = 'content';
   const header = 'header';
