+++ packages/webapp/src/shared/components/auth/twoFactorAuthForm/twoFactorAuthForm.component.tsx
@@ -1,7 +1,7 @@
 import { useMutation } from '@apollo/client';
 import { useCommonQuery } from '@sb/webapp-api-client/providers';
 import { Button } from '@sb/webapp-core/components/buttons';
-import { Dialog, DialogContent } from '@sb/webapp-core/components/dialog';
+import { Dialog, DialogContent } from '@sb/webapp-core/components/ui/dialog';
 import { useOpenState } from '@sb/webapp-core/hooks';
 import { trackEvent } from '@sb/webapp-core/services/analytics';
 import { useToast } from '@sb/webapp-core/toast/useToast';
